requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
sqlmodel==0.0.21
SQLAlchemy==2.0.35
# psycopg2-binary pinned for Docker/Linux; omit for local Windows dev
python-dotenv==1.0.1
httpx==0.27.2
feedparser==6.0.11
apscheduler==3.10.4
redis==5.0.8
celery==5.4.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
pydantic==2.9.2
email-validator==2.2.0
scikit-learn==1.5.2
numpy==2.1.1
pandas==2.2.3
orjson==3.10.7
tenacity==9.0.0
